Reuse predictor class validations for tasks
