+++ pkg/cortex/serve/cortex_internal/lib/api/predictor.py
@@ -25,6 +25,10 @@ import dill
 from datadog import DogStatsd
 
 from cortex_internal.lib import util
+from cortex_internal.lib.api.validations import (
+    validate_class_impl,
+    validate_python_predictor_with_models, _are_models_specified,
+)
 from cortex_internal.lib.client.onnx import ONNXClient
 from cortex_internal.lib.client.python import PythonClient
 from cortex_internal.lib.client.tensorflow import TensorFlowClient
@@ -50,6 +54,79 @@ from cortex_internal.lib.type import (
 )
 
 logger = configure_logger("cortex", os.environ["CORTEX_LOG_CONFIG_FILE"])
+
+PYTHON_CLASS_VALIDATION = {
+    "required": [
+        {
+            "name": "__init__",
+            "required_args": ["self", "config"],
+            "optional_args": ["job_spec", "python_client", "metrics_client"],
+        },
+        {
+            "name": "predict",
+            "required_args": ["self"],
+            "optional_args": ["payload", "query_params", "headers", "batch_id"],
+        },
+    ],
+    "optional": [
+        {"name": "on_job_complete", "required_args": ["self"]},
+        {
+            "name": "post_predict",
+            "required_args": ["self"],
+            "optional_args": ["response", "payload", "query_params", "headers"],
+        },
+        {
+            "name": "load_model",
+            "required_args": ["self", "model_path"],
+        },
+    ],
+}
+
+TENSORFLOW_CLASS_VALIDATION = {
+    "required": [
+        {
+            "name": "__init__",
+            "required_args": ["self", "tensorflow_client", "config"],
+            "optional_args": ["job_spec", "metrics_client"],
+        },
+        {
+            "name": "predict",
+            "required_args": ["self"],
+            "optional_args": ["payload", "query_params", "headers", "batch_id"],
+        },
+    ],
+    "optional": [
+        {"name": "on_job_complete", "required_args": ["self"]},
+        {
+            "name": "post_predict",
+            "required_args": ["self"],
+            "optional_args": ["response", "payload", "query_params", "headers"],
+        },
+    ],
+}
+
+ONNX_CLASS_VALIDATION = {
+    "required": [
+        {
+            "name": "__init__",
+            "required_args": ["self", "onnx_client", "config"],
+            "optional_args": ["job_spec", "metrics_client"],
+        },
+        {
+            "name": "predict",
+            "required_args": ["self"],
+            "optional_args": ["payload", "query_params", "headers", "batch_id"],
+        },
+    ],
+    "optional": [
+        {"name": "on_job_complete", "required_args": ["self"]},
+        {
+            "name": "post_predict",
+            "required_args": ["self"],
+            "optional_args": ["response", "payload", "query_params", "headers"],
+        },
+    ],
+}
 
 
 class Predictor:
@@ -249,9 +326,9 @@ class Predictor:
             raise
 
         try:
-            _validate_impl(predictor_class, validations, self.api_spec)
+            validate_class_impl(predictor_class, validations)
             if self.type == PythonPredictorType:
-                _validate_python_predictor_with_models(predictor_class, self.api_spec)
+                validate_python_predictor_with_models(predictor_class, self.api_spec)
         except (UserException, Exception) as e:
             e.wrap("error in " + self.path)
             raise
@@ -301,191 +378,6 @@ class Predictor:
             cron.stop()
         for cron in self.crons:
             cron.join()
-
-
-def _are_models_specified(api_spec: Dict) -> bool:
-    """
-    Checks if models have been specified in the API spec (cortex.yaml).
-
-    Args:
-        api_spec: API configuration.
-    """
-    predictor_type = predictor_type_from_api_spec(api_spec)
-
-    if predictor_type == PythonPredictorType and api_spec["predictor"]["multi_model_reloading"]:
-        models = api_spec["predictor"]["multi_model_reloading"]
-    elif predictor_type != PythonPredictorType:
-        models = api_spec["predictor"]["models"]
-    else:
-        return False
-
-    return models is not None
-
-
-PYTHON_CLASS_VALIDATION = {
-    "required": [
-        {
-            "name": "__init__",
-            "required_args": ["self", "config"],
-            "optional_args": ["job_spec", "python_client", "metrics_client"],
-        },
-        {
-            "name": "predict",
-            "required_args": ["self"],
-            "optional_args": ["payload", "query_params", "headers", "batch_id"],
-        },
-    ],
-    "optional": [
-        {"name": "on_job_complete", "required_args": ["self"]},
-        {
-            "name": "post_predict",
-            "required_args": ["self"],
-            "optional_args": ["response", "payload", "query_params", "headers"],
-        },
-        {
-            "name": "load_model",
-            "required_args": ["self", "model_path"],
-        },
-    ],
-}
-
-TENSORFLOW_CLASS_VALIDATION = {
-    "required": [
-        {
-            "name": "__init__",
-            "required_args": ["self", "tensorflow_client", "config"],
-            "optional_args": ["job_spec", "metrics_client"],
-        },
-        {
-            "name": "predict",
-            "required_args": ["self"],
-            "optional_args": ["payload", "query_params", "headers", "batch_id"],
-        },
-    ],
-    "optional": [
-        {"name": "on_job_complete", "required_args": ["self"]},
-        {
-            "name": "post_predict",
-            "required_args": ["self"],
-            "optional_args": ["response", "payload", "query_params", "headers"],
-        },
-    ],
-}
-
-ONNX_CLASS_VALIDATION = {
-    "required": [
-        {
-            "name": "__init__",
-            "required_args": ["self", "onnx_client", "config"],
-            "optional_args": ["job_spec", "metrics_client"],
-        },
-        {
-            "name": "predict",
-            "required_args": ["self"],
-            "optional_args": ["payload", "query_params", "headers", "batch_id"],
-        },
-    ],
-    "optional": [
-        {"name": "on_job_complete", "required_args": ["self"]},
-        {
-            "name": "post_predict",
-            "required_args": ["self"],
-            "optional_args": ["response", "payload", "query_params", "headers"],
-        },
-    ],
-}
-
-
-def _validate_impl(impl, impl_req, api_spec):
-    for optional_func_signature in impl_req.get("optional", []):
-        _validate_optional_fn_args(impl, optional_func_signature)
-
-    for required_func_signature in impl_req.get("required", []):
-        _validate_required_fn_args(impl, required_func_signature)
-
-
-def _validate_optional_fn_args(impl, func_signature):
-    if getattr(impl, func_signature["name"], None):
-        _validate_required_fn_args(impl, func_signature)
-
-
-def _validate_required_fn_args(impl, func_signature):
-    target_class_name = impl.__name__
-
-    fn = getattr(impl, func_signature["name"], None)
-    if not fn:
-        raise UserException(
-            f"class {target_class_name}",
-            f'required method "{func_signature["name"]}" is not defined',
-        )
-
-    if not callable(fn):
-        raise UserException(
-            f"class {target_class_name}",
-            f'"{func_signature["name"]}" is defined, but is not a method',
-        )
-
-    required_args = func_signature.get("required_args", [])
-    optional_args = func_signature.get("optional_args", [])
-
-    argspec = inspect.getfullargspec(fn)
-    fn_str = f'{func_signature["name"]}({", ".join(argspec.args)})'
-
-    for arg_name in required_args:
-        if arg_name not in argspec.args:
-            raise UserException(
-                f"class {target_class_name}",
-                f'invalid signature for method "{fn_str}"',
-                f'"{arg_name}" is a required argument, but was not provided',
-            )
-
-        if arg_name == "self":
-            if argspec.args[0] != "self":
-                raise UserException(
-                    f"class {target_class_name}",
-                    f'invalid signature for method "{fn_str}"',
-                    f'"self" must be the first argument',
-                )
-
-    seen_args = []
-    for arg_name in argspec.args:
-        if arg_name not in required_args and arg_name not in optional_args:
-            raise UserException(
-                f"class {target_class_name}",
-                f'invalid signature for method "{fn_str}"',
-                f'"{arg_name}" is not a supported argument',
-            )
-
-        if arg_name in seen_args:
-            raise UserException(
-                f"class {target_class_name}",
-                f'invalid signature for method "{fn_str}"',
-                f'"{arg_name}" is duplicated',
-            )
-
-        seen_args.append(arg_name)
-
-
-def _validate_python_predictor_with_models(impl, api_spec):
-    target_class_name = impl.__name__
-
-    if _are_models_specified(api_spec):
-        constructor = getattr(impl, "__init__")
-        constructor_arg_spec = inspect.getfullargspec(constructor)
-        if "python_client" not in constructor_arg_spec.args:
-            raise UserException(
-                f"class {target_class_name}",
-                f'invalid signature for method "__init__"',
-                f'"python_client" is a required argument, but was not provided',
-                f'when the python predictor type is used and models are specified in the api spec, adding the "python_client" argument is required',
-            )
-
-        if getattr(impl, "load_model", None) is None:
-            raise UserException(
-                f"class {target_class_name}",
-                f'required method "load_model" is not defined',
-                f'when the python predictor type is used and models are specified in the api spec, adding the "load_model" method is required',
-            )
 
 
 def model_downloader(
